Fixes #15090: Run deletion protection rules prior to enqueueing events
